requirements.txt
alembic==1.10.2
aniso8601==9.0.1
anyio==3.6.2
autocorrect==2.6.1
autopep8==1.6.0
backoff==2.2.1
blis==0.7.9
boto3==1.26.97
botocore==1.29.97
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==14.0
croniter==1.3.8
cymem==2.0.7
dagit==1.2.1
dagster==1.2.1
dagster-aws==0.18.1
dagster-graphql==1.2.1
dagster-pandas==0.18.1
dagster-postgres==0.18.1
docstring-parser==0.15
flake8==4.0.1
fsspec==2023.3.0
gql==3.4.0
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
grpcio==1.47.5
grpcio-health-checking==1.47.5
h11==0.14.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
langcodes==3.3.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.6.1
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pendulum==2.1.2
preshed==3.0.8
protobuf==4.22.1
psycopg2-binary==2.9.5
pyarrow==11.0.0
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
srsly==2.4.6
starlette==0.26.1
StrEnum==0.4.10
tabulate==0.9.0
thinc==8.0.17
toml==0.10.2
tomli==2.0.1
toposort==1.10
tqdm==4.65.0
typer==0.4.2
typing_extensions==4.5.0
universal_pathlib==0.0.22
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
wasabi==0.10.1
watchdog==3.0.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
